Added ActUpon methods for struct and reference types
Implemented two new ActUpon methods in the Result class. These methods allow actions to be performed on struct and reference type public properties of entities. One method handles reference types, while the other one is for struct types. Both methods take a selector, an action, and an argument as parameters. They invoke the action only if the Result instance is valid.