Feature: Added Short-Circuiting Mechanism to                RuleSequence
This commit introduces a short-circuiting mechanism to the `RuleSequence` class. This mechanism allows the evaluation of rules to be disrupted when a rule that enables short-circuiting fails. The `RuleSequence` constructor has been refactored to separate the rules into two arrays based on the `enablesShortCircuiting` flag. Rules with `enablesShortCircuiting` set to `true` are added to the `ShortCircuitRules` array, and rules with `enablesShortCircuiting` set to `false` are added to the `Rules` array. This allows for more efficient and flexible rule evaluation.