Added EvaluationActionProvider, DirectAction, and IndirectAction Classes
These classes are designed to provide evaluation actions based on the type of `DomainRule`.

These classes allow to handle the `TProperty` type being unknown when calling an external function. They also help us infer the type and avoid the use of reflection.

An important aspect of these classes is that the `Singular` and `Sequencial` functions in both `DirectAction<TSubject>` and `IndirectAction<TSubject, TProperty>` classes are created only once throughout the lifetime of the application. This can improve performance.